Changed textures to better resemble stone
Also enabled parallax mapping (should make textures more 3D), but it doesn't seem to do much with the current heightmaps (might have to change scaling factor in vmt).

Currently evaluating two textures, one with a SSBump and the other with just a normal. SSBump (all of the grey rock textures to the right of the door) looks better and should run faster than normal mode (right of the door)